refactor: auth and spotlight context to be own hooks
