+++ apps/web/src/components/auth/PasswordResetForm.tsx
@@ -1,4 +1,3 @@
-import { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useMutation } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
@@ -6,7 +5,7 @@ import * as Sentry from '@sentry/react';
 import { showNotification } from '@mantine/notifications';
 import { passwordConstraints } from '@novu/shared';
 
-import { AuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../store/authContext';
 import { api } from '../../api/api.client';
 import { PasswordInput, Button, colors, Text } from '../../design-system';
 import { PasswordRequirementPopover } from './PasswordRequirementPopover';
@@ -17,7 +16,7 @@ type Props = {
 };
 
 export function PasswordResetForm({ token }: Props) {
-  const { setToken } = useContext(AuthContext);
+  const { setToken } = useAuthContext();
 
   const navigate = useNavigate();
   const { isLoading, mutateAsync, isError, error } = useMutation<
