+++ apps/web/src/components/utils/Spotlight.tsx
@@ -1,13 +1,13 @@
 import { SpotlightProvider } from '@mantine/spotlight';
-import { useContext, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Activity, Bolt, Box, Settings, Repeat, Team, Brand, Chat } from '../../design-system/icons';
-import { SpotlightContext } from '../../store/spotlightContext';
+import { useSpotlightContext } from '../../store/spotlightContext';
 import { ROUTES } from '../../constants/routes.enum';
 
 export const SpotLight = ({ children }) => {
   const navigate = useNavigate();
-  const { items, addItem } = useContext(SpotlightContext);
+  const { items, addItem } = useSpotlightContext();
 
   useEffect(() => {
     addItem([
