+++ apps/web/src/pages/auth/LoginPage.tsx
@@ -1,9 +1,9 @@
-import { useContext, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import jwtDecode from 'jwt-decode';
 import { IJwtPayload } from '@novu/shared';
 
-import { AuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../store/authContext';
 import { LoginForm } from '../../components/auth/LoginForm';
 import AuthLayout from '../../components/layout/components/AuthLayout';
 import AuthContainer from '../../components/layout/components/AuthContainer';
@@ -17,7 +17,7 @@ import { ROUTES } from '../../constants/routes.enum';
 
 export default function LoginPage() {
   useBlueprint();
-  const { setToken, token: oldToken } = useContext(AuthContext);
+  const { setToken, token: oldToken } = useAuthContext();
   const segment = useSegment();
   const navigate = useNavigate();
   const [params] = useSearchParams();
