+++ apps/web/src/pages/brand/BrandPage.tsx
@@ -1,18 +1,17 @@
-import { useContext } from 'react';
 import { Container } from '@mantine/core';
 
 import PageMeta from '../../components/layout/components/PageMeta';
 import PageHeader from '../../components/layout/components/PageHeader';
 import PageContainer from '../../components/layout/components/PageContainer';
 import { Tabs } from '../../design-system';
-import { AuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../store/authContext';
 import { BrandingForm, LayoutsListPage } from './tabs';
 
 const BRANDING = 'Assets';
 const LAYOUT = 'Layouts';
 
 export function BrandPage() {
-  const { currentOrganization } = useContext(AuthContext);
+  const { currentOrganization } = useAuthContext();
 
   const menuTabs = [
     {
