+++ apps/web/src/pages/settings/SettingsPage.tsx
@@ -1,4 +1,3 @@
-import { useContext } from 'react';
 import { Container } from '@mantine/core';
 
 import PageMeta from '../../components/layout/components/PageMeta';
@@ -6,7 +5,7 @@ import PageHeader from '../../components/layout/components/PageHeader';
 import PageContainer from '../../components/layout/components/PageContainer';
 import { Tabs } from '../../design-system';
 import { ApiKeysCard, InAppCenterCard } from './tabs';
-import { AuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../store/authContext';
 import { EmailSettings } from './tabs/EmailSettings';
 
 enum MenuTitleEnum {
@@ -16,7 +15,7 @@ enum MenuTitleEnum {
 }
 
 export function SettingsPage() {
-  const { currentOrganization } = useContext(AuthContext);
+  const { currentOrganization } = useAuthContext();
   const selfHosted = process.env.REACT_APP_DOCKER_HOSTED_ENV === 'true';
 
   let menuTabs = [
