+++ apps/web/src/pages/templates/editor/EmailIntegrationInfo.tsx
@@ -1,7 +1,6 @@
-import { useContext } from 'react';
 import { useIntegrationLimit } from '../../../api/hooks/integrations/useIntegrationLimit';
 import { ChannelTypeEnum } from '@novu/shared';
-import { AuthContext } from '../../../store/authContext';
+import { useAuthContext } from '../../../store/authContext';
 
 export const EmailIntegrationInfo = ({
   integration,
@@ -16,7 +15,7 @@ export const EmailIntegrationInfo = ({
   field: 'from' | 'senderName';
 }) => {
   const { enabled, loading } = useIntegrationLimit(ChannelTypeEnum.EMAIL);
-  const { currentOrganization } = useContext(AuthContext);
+  const { currentOrganization } = useAuthContext();
 
   if (integration) {
     return <>{integration?.credentials[field]}</>;
